IN-496 stash useful functions to helpers
